Clarify duplicates in list
See GiveWell page for example:
https://orgwatch.issarice.com/?organization=GiveWell
(e.g. search for "taylor smith, taylor smith").

This is just a quick fix to let people know that this is the intended
behavior for now, not some bug.  However, the duplicates thing is
confusing enough that I do think eventually it would be better to just
fix how the data is displayed.